Added dhcpcd options back in, and set interfaces to manual.
Apparently raspbian (or all newer debians) use some funky combination of /etc/network/interfaces and dhcpcd. If you specify a network interface as "dhcp", it will use dhclient and not dhcpcd. dhcpcd just tries to configure any interface unless you tell it not to. It's the one for which we can specify shorter timeout/retry intervals, so let's use that one.